Good or bad

+++ problem3.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
 using namespace std;
 int main() {
-    int c=0;
+    double c=0;
     cin>>c;
     double f=c*9/5+32;
     cout<<f<<endl;
